requirements.txt
httpx==0.27.2
openai>=1.52.0
python-dotenv>=1.0.0
streamlit>=1.32.0